Font get_width() returns now float, to avoid rounding errors

+++ src/font.hpp
@@ -38,8 +38,8 @@ public:
   void render(Origin origin, int x, int y, const std::string& text, Framebuffer& fb);
 
   int  get_height() const;
-  int  get_width(uint32_t unicode) const;
-  int  get_width(const std::string& text) const;
+  float get_width(uint32_t unicode) const;
+  float get_width(const std::string& text) const;
   Size get_size(const std::string& text) const;
   Rect bounding_rect(int , int, const std::string& str) const;
